Update assertionStore; Update to folder and add models
