+++ src/components/Auditor/AuditorImport.svelte
@@ -6,7 +6,7 @@
 
 <script>
   import { getContext } from 'svelte';
-  import { importAssertions } from '@app/stores/earl/assertionStore.js';
+  import { importAssertions } from '@app/stores/earl/assertionStore/index.js';
   import File, { readFile } from '@app/components/formcomponents/File.svelte';
 
   const { translate } = getContext('app');
